ADD: ProofFlow now uses MV format by default.
Outputconfigs had to be moved because otherwise there was a dependency loop.
