+++ src/ProofFlow/parser/parsers.ts
@@ -1,13 +1,9 @@
-import { OutputConfig } from "../editor/ProofFlowDocument";
 import { SimpleParser } from "./parser";
 
 export {
   CoqParser,
   CoqMDParser,
   LeanParser,
-  CoqOutput,
-  CoqMDOutput,
-  LeanOutput,
 };
 
 const CoqParser = new SimpleParser({
@@ -30,30 +26,3 @@ const LeanParser = new SimpleParser({
   collapsible: [/:::collapsible\n(?:# (.*?)\n)?/, /:::\n/],
   input: [/:::input\n/, /:::\n/],
 });
-
-const CoqOutput: OutputConfig = {
-  text: ["(**", "*)"],
-  code: ["", ""],
-  math: ["$$", "$$"],
-  collapsible: ["<hint>", "</hint>"],
-  collapsibletitle: ['<hint title="TITLE">\n', "\n</hint>"],
-  input: ["<input-area>\n", "\n</input-area>"],
-};
-
-const CoqMDOutput: OutputConfig = {
-  text: ["", ""],
-  code: ["\n```coq\n", "\n```"],
-  math: ["$$", "$$"],
-  collapsible: ["<hint>", "\n</hint>"],
-  collapsibletitle: ['<hint title="TITLE">', "\n</hint>"],
-  input: ["<input-area>", "\n</input-area>"],
-};
-
-const LeanOutput: OutputConfig = {
-  text: ["", ""],
-  code: ["```lean\n", "```\n"],
-  math: [":::math\n", ":::\n"],
-  collapsible: [":::collapsible\n", ":::\n"],
-  collapsibletitle: [":::collapsible\n# TITLE\n", ":::\n"],
-  input: [":::input\n", ":::\n"],
-};
